requirements.txt
requests>=2.31.0
netCDF4>=1.6.4
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0 